requirements.txt
googlemaps
djangorestframework
markdown
pyyaml